refactor: consolidate utils.py into _utils.py
Move parse_version from utils.py to _utils.py and remove utils.py.
